refactor: clean up i18n store helpers
