+++ packages/react-core/src/i18n-store/storeTypes.ts
@@ -9,7 +9,6 @@ import type {
 
 export type Unsubscribe = () => void;
 export type StoreListener = () => void;
-export type ListenerSet = Set<StoreListener>;
 
 // ----- Lookups ----- //
