+++ packages/react-core/src/i18n-store/utils/subscriptions.ts
@@ -1,0 +1,11 @@
+import type { Unsubscribe } from '../storeTypes';
+
+export function subscribeToSet<T>(
+  listenerSet: Set<T>,
+  listener: T
+): Unsubscribe {
+  listenerSet.add(listener);
+  return () => {
+    listenerSet.delete(listener);
+  };
+}
